camera_v2: Restore original camera_v2 code

+++ drivers/media/platform/msm/camera_v2/isp/msm_buf_mgr.c
@@ -145,7 +145,7 @@ static int msm_isp_prepare_v4l2_buf(struct msm_isp_buf_mgr *buf_mgr,
 		}
 		mapped_info->paddr += v4l2_buf->m.planes[i].data_offset;
 		CDBG("%s: plane: %d addr:%lu\n",
-			__func__, i, mapped_info->paddr);
+			__func__, i, (unsigned long int)(mapped_info->paddr));
 
 		buf_pending = kzalloc(sizeof(struct buffer_cmd), GFP_ATOMIC);
 		if (!buf_pending) {
